+++ initialize.sh
